Moved PartyTablesTitle to presentational
